fix: move backfill script to sms folder
