+++ src/inngest/functions/sms/goodbyeSMSCommunicationJourney.ts
@@ -6,7 +6,7 @@ import { onScriptFailure } from '@/inngest/onScriptFailure'
 import { sendSMS } from '@/utils/server/sms'
 import * as messages from '@/utils/server/sms/messages'
 
-import { createCommunication, createCommunicationJourneys } from './utils/communicationJourney'
+import { createCommunication, createCommunicationJourneys } from './utils'
 
 export const GOODBYE_SMS_COMMUNICATION_JOURNEY_INNGEST_EVENT_NAME =
   'app/user.communication/goodbye.sms'
