+++ src/inngest/functions/sms/index.ts
@@ -1,3 +1,4 @@
+export { backfillPhoneNumberValidation } from './backfillPhoneNumberValidation'
 export { backfillSMSStatusField } from './backfillSMSStatusField'
 export { bulkSMSCommunicationJourney } from './bulkSMSCommunicationJourney'
 export { goodbyeSMSCommunicationJourney } from './goodbyeSMSCommunicationJourney'
